Correct population condition processing
The existing scheme would only work correctly for the works table.
The new scheme evaluates the condition against the joined tables
and then populates the tables based on the records identified as
included in the result.  The new scheme also allows using an arbitrary
SQL expression for the condition, rather than the table:condition
format.

The new scheme is four times more expensive than the existing one.
This happens because of the multiple times the virtual tables are
scanned.

It can be optimized as follows:
- Create temporary table copies for the virtual tables
- Create temporary tables in memory
- Use an APSW execution tracer to populate the temporary tables
  only with the columns required for executing the query
  https://rogerbinns.github.io/apsw/tips.html#parsing-sql