Refactor error handling
All fatal errors are now raised as an Alexandria3kError or
an Alexandria3kInternalError.

The CLI catches the Alexandria3kError and reports it through its error
message.  Internal errors are not caught by the CLI, and are reported
as stack traces.  The "stacktrace" debug option can be set to display a
stack trace, as was the case with the (now removed) "exception" debug
option.

The API raises Alexandria3kError rather than terminating its program
with an error message and (optionally) a stack trace.