Support apsw 3.48
Create the virtual tables in a temporary rather than in-memory
database.

Accessing shared memory databases by name with the existing
scheme seems to be deprecated.  When a file-based databased is
attached to an in-memory database with the new VFS memdb scheme,
the file-based database is not written to, and hence database
population of file-based databases via an in-memory database
containing the virtual tables fails.

See https://github.com/rogerbinns/apsw/issues/553#issuecomment-2604313098